// web/live-anchor/src/views/message/utils/types.ts
export enum MessageCommand {
  ping = 13,
  pingResponse = 13,
  receive = 11,
  send = 11,
  sendResponse = 12,
  login = 5,
  loginResponse = 6,
  logout = 22,
  logoutResponse = 22,
  joinGroup = 7,
  joinGroupResponse = 9,
  closeLive = 26,
  openLive = 25,
  leaveGroup = 21,
  leaveGroupResponse = 10,
  read = 23,
  matchChange = 30,
  readDotAll = 29,
  anchorReadMessage = 31,
  banned = 32,
  unBanned = 33,
  feedback = 28,
  liveStartNotify = 34,
  updateUrl = 27
}
export enum MessageType {
  text = 0,
  image = 1,
  voice = 2,
  video = 3,
  music = 4,
  news = 5
}
type MessageModel<T> = {
  command: MessageCommand; // 11
  code: number; //
  success: number; //
  data: T;
};
// ...
export type MessageReceiveData = MessageModel<{
  anchorId: string; //（字符串）: 当前聊天主播的id【私聊独有，不管是发送还是接收】
  chatType: number; //聊天类型，群聊：1， 私聊：2
  cmd: number;
  content: string; //消息内容
  createTime?: number; //(长整型): 消息时间
  from: string; //（字符串）: 发送者ID
  groupId?: string; //: 群聊ID【群聊独有】
  id: string; //消息ID
  msgType: MessageType; //消息类型，文字：0， 图片：1
  sendId?: string; //（字符串）:  消息发送标识符
  to: string; //（字符串）: 接收者ID【私聊独有】
  toAvatar: string; // 接收者头像（主播、运营、助手均显示主播头像）
  toNickName: string; // 接收者昵称（主播、运营、助手均显示主播昵称）
  fromAvatar: string; // 发送者头像（主播、运营、助手均显示主播头像）
  fromNickName: string; // 发送者昵称（主播、运营、助手均显示主播昵称）
  level: string; // level: 用户等级（发送者为用户时必填）
  identityType: 0 | 1 | 2 | 3; // 发送者类型：0-APP普通用户，1-LIVE端主播，2-LIVE端助手，3-LIVE端运营
}>;
// ...
export enum MessageCardType {
  SYSTEM = "system",
  MESSAGE = "message"
}
export type MessageCardData = {
  id: string;
  anchorId: string;
  sendUserId: string;
  name: string;
  cover: string;
  message: string;
  readDot: number;
  createTime: string;
  createTimeText: string;
  type: MessageCardType;
  msgType: MessageType;
};